third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-reconnect-during-with-uuid.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let device, fake_peripheral, service;
  return getHealthThermometerService()
    .then(_ => ({device, fake_peripheral, service} = _))
    .then(() => {
      let promise = assert_promise_rejects_with_message(service.getCharacteristics('heart_rate_measurement'), new DOMException(
          'GATT Server is disconnected. Cannot retrieve characteristics. ' +
          '(Re)connect first with `device.gatt.connect`.',
          'NetworkError'));
      device.gatt.disconnect();
      return fake_peripheral.setNextGATTConnectionResponse({code: HCI_SUCCESS})
          .then(() => device.gatt.connect())
          .then(() => promise);
    });
}, 'disconnect() and connect() called during getCharacteristics. Reject with ' +
   'NetworkError.');

</script>
